--- src/main/java/Parser.java
@@ -1,18 +1,0 @@
-import java.util.Scanner;
-public class Parser {
-    private final Scanner scanner;
-
-    Parser() {
-        this.scanner = new Scanner(System.in);
-    }
-
-    String getInput() {
-        return this.scanner.nextLine();
-
-    }
-
-    void closeParser() {
-        scanner.close();
-    }
-
-}
