+++ src/main/java/duke/Duke.java
@@ -1,4 +1,4 @@
-
+package duke;
 public class Duke {
 
     private final CommandProcessor commandProcessor;
