+++ src/main/java/duke/Parser.java
@@ -1,0 +1,20 @@
+package duke;
+
+import java.util.Scanner;
+public class Parser {
+    private final Scanner scanner;
+
+    Parser() {
+        this.scanner = new Scanner(System.in);
+    }
+
+    String getInput() {
+        return this.scanner.nextLine();
+
+    }
+
+    void closeParser() {
+        scanner.close();
+    }
+
+}
