+++ src/test/java/duke/EventTest.java
@@ -1,0 +1,17 @@
+package duke;
+
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class EventTest {
+    @Test
+    public void testStorageText() {
+        try {
+            assertEquals("[E][ ] trying books /from 2015-07-09 /to 2017-01-10"
+                    , new Event("trying books /from 2015-07-09 /to 2017-01-10").storageText());
+        } catch (DukeException e) {
+            System.out.println(e.getMessage());
+        }
+    }
+
+}
